updated use of log file

+++ root/usr/local/bin/save_to_backup.sh
@@ -79,7 +79,7 @@ function save_to_backup() {
 
   max_file_mod_time=$(date --date="${new_tar_date}") 
   min_file_mod_time=$(date --date="${last_tar_date}")
-  logf "${log_tag}" "New tar files will only include files modified between ${min_file_mod_time} and ${max_file_mod_time}"
+  logf "${log_tag}" "New tar files will only include files modified between ${min_file_mod_time} and ${max_file_mod_time}" "${log_file}"
 
   new_tar_file_no_ext="${tar_filename_start}_${last_tar_date}_to_${new_tar_date}"
   logf "${log_tag}" "New tar files name prefix ${new_tar_file_no_ext}" "${log_file}"
@@ -110,7 +110,10 @@ function save_to_backup() {
 
     tar --create -z --file="${split_new_tar_file}" --files-from="${split_list_file}"
 
-    stat "${split_new_tar_file}" "${log_file}"
+    logf "${log_tag}"  "Tar file stat:" "${log_file}"
+    tar_info=$(stat "${split_new_tar_file}")
+    echo "${tar_info}"
+    echo "${tar_info}" >> "${log_file}"
 
     if  gzip -v -t "${split_new_tar_file}" 2>  /dev/null; then
         logf "${log_tag}"  "tar gzip compression tested ok, moving ${split_new_tar_file} to dir ${local_backup_dir}" "${log_file}"
@@ -126,7 +129,7 @@ function save_to_backup() {
 
         #break
     fi
-    logf "${log_tag}"  "$ ****** Finished image Libary tar file rebuild for ${split_new_tar_file} ******"  "${log_file}"
+    logf "${log_tag}"  "****** Finished image Libary tar file rebuild for ${split_new_tar_file} ******"  "${log_file}"
     echo "" >> "${log_file}"
     echo "" >> "${log_file}"
   done
@@ -134,7 +137,11 @@ function save_to_backup() {
 
   logf "${log_tag}" "****** Syncing backup tar files to rclone remote: "${rclone_remote}:${rclone_path}" ******" "${log_file}"
  
-  cp "${log_file}" "${config_dir}/"
+  if [ -f "${log_file}" ]; then
+    cp "${log_file}" "${config_dir}/"
+  else
+    errorf "${log_tag}"  "****** Unable to find log file ${log_file} to copy to ${config_dir} ******" 
+  fi
 
 
 # Only tars made in this session
@@ -142,7 +149,7 @@ function save_to_backup() {
     --config "${restore_rclone_config}" \
     --stats 5m \
     --progress \
-    --log-file "${log_file}" \
+    --log-file "${log_file}.sync.log" \
     --filter "+ ${new_tar_file_no_ext}_*.tar.gz" \
     --filter "- *"
 
@@ -155,7 +162,14 @@ function save_to_backup() {
 
   logf "${log_tag}" "completed save_to_backup" "${log_file}"
 
-  mv "${log_file}" "${config_dir}/"
+
+  if [ -f "${log_file}" ]; then
+    mv "${log_file}" "${config_dir}/"
+    mv "${log_file}.sync.log"" "${config_dir}/"
+  else
+    errorf "${log_tag}"  "****** Unable to find log file ${log_file} to move to ${config_dir} ******" 
+  fi
+  
 }
 
 # # Set the target directory
